+++ RELEASENOTES.md
@@ -2,7 +2,23 @@
 # React Components: design-system-react
 # Release notes
 
+## Release 0.1.0
+
+**BREAKING CHANGES**
+- Rename ContextBar to GlobalNavigationBar
+
+**MAJOR CHANGES**
+- Add the Global Header component
+- Add a Search component
+- Updates BreadCrumb to the latest markup
+
+**OTHER**
+- Begin counting versions from 0.1.0
+- Add buffered option to DataTable
+
+
 ## Release 0.0.35
+
 **MAJOR CHANGES**
 - Reversed Flippable prop due to bug with Tether.js
 - Use Popover component in Datepicker
@@ -10,6 +26,7 @@
 
 
 ## Release 0.0.34
+
 **BREAKING CHANGES**
 - Flippable prop on Timepicker and Datepicker used to default to true. Now, it defaults to false and you must explicitly add it.
 
@@ -24,8 +41,10 @@
 - Remove unused Popover class on Picklist
 - Additional code cleanup
 
+
 ## Release 0.0.32
 - Added containerClassName prop to Spinner
+
 
 ## Release 0.0.31
 
@@ -39,6 +58,7 @@
 - Remove isRequired for iconCategory in ButtonGroup
 - Additional bug fixes and code cleanup
 
+
 ## Release 0.0.29
 
 **MAJOR CHANGES**
@@ -46,6 +66,7 @@
 
 **SITE AND DOCS**
 - Moved site to external repo
+
 
 ## Release 0.0.28
 
@@ -56,6 +77,7 @@
   - SLDSGrid
   - SLDSPageHeader
 - Lookup component no longer requires icons for menu items
+
 
 ## Release 0.0.27